style: move page header styling to App
