+++ client/src/components/PageTitleBar.tsx
@@ -1,6 +1,5 @@
 import { PageHeader, PageHeaderProps, Typography } from "antd";
 import React from "react";
-import "./PageTitleBar.css";
 
 const { Text } = Typography;
