--- client/src/components/PageTitleBar/PageTitleBar.css
@@ -1,3 +1,0 @@
-.pageTitleBar {
-  padding: 0;
-}
