+++ client/src/pages/DatasetDetails/DatasetDetails.tsx
@@ -2,7 +2,6 @@ import { PlusOutlined } from "@ant-design/icons";
 import { useQuery } from "@apollo/client";
 import { Button } from "antd";
 import { useTranslation } from "react-i18next";
-import React from "react";
 import { useHistory, useParams } from "react-router-dom";
 import {
   GetDataset,
@@ -11,7 +10,7 @@ import {
 import { GET_DATASET } from "../../graphql/__queries__/GetDataset.gql";
 import { DatasetDetailsRecordsTable } from "./DatasetDetailsRecordsTable";
 import { DatasetDetailsScoreCard } from "./DatasetDetailsScoreCard";
-import { PageTitleBar } from "../../components/PageTitleBar/PageTitleBar";
+import { PageTitleBar } from "../../components/PageTitleBar";
 
 interface RouteParams {
   datasetId: string;
