+++ client/src/pages/DatasetDetails/DatasetDetailsScoreCard.css
@@ -1,4 +1,3 @@
 .dataset-details_statistics {
   margin-bottom: 1rem;
-  margin-top: 1rem;
 }
